requirements.txt
Flask
beautifulsoup4
sounddevice
openai-whisper
nltk
numpy
torch

